Remove VOLUME
The end-user should be left to determine if they want to bake configs into the image or use a volume. 

While I understand that the $JENKINS_HOME variable can be used to override the directory in child images, that is not conventional in official images on Docker Hub. Official builds are often used immutable building blocks for other containers and those containers are where the determination of storage should be made. Even if a child image is not made, a user can easily specify a volume on the command-line for storing state outside of a container.