Vary the shallow settings in JENKINS-31393 test jobs
The test jobs were failing on initial startup because they were trying to
build a reference that was not the HEAD reference, but were only cloning
shallow to depth 1.  Since the shallow clone at depth 1 had only the
most recent commit, it could not find the SHA1 that was being attempted.