Upgrade plugins if needed
Each plugin installed through the docker image is now tracked using a
marker file containing the version that was installed through the image.

If later, the container is started from a docker image containing newer
plugins, existing plugins will be upgraded if they haven't been upgraded
in the mean time and if the new docker image provides a new version of
them.