Add buildresult trigger plugin
JENKINS-06203, JENKINS-14798, and JENKINS-29482 jobs each need a job which
writes to their repository so that they can then perform their checks.
Without the commit to the repository, the job (correctly) fails.

This plugin will allow paired jobs which detect the failure of the
original job, push a commit to the necessary repository, and trigger
polling of that repository.