Change filename for number of executors example
When trying to run that script locally, it does not work (tested in `jenkins:2.32.2-alpine`). 
There is already a file named `executors.groovy` and so if you try to add the file, it won't actually override what is in the settings. When I changed the file name, it was able to work.