Resolve plugins from version-specific update center
Reads Jenkins version from the  manifest and checks if an update center
for this version is available. If this is the case, latest plugin
versions are resolved from this update center. This ensures that
latest plugin versions are always compatible to the LTS version if
an LTS version is used.

Fixes #494