Drop deprecated java base image for openjdk
This base image is officially deprecated in favor of the openjdk image,
and will receive no further updates after 2016-12-31 (Dec 31, 2016).
Making the change also fixes the breakage due to the mixlib gem requiring
ruby greater than the 2.1 version supplied by the deprecated java base
as the openjdk image comes with 2.3.