Fix failure loop if installing plugins fails
`install-plugins.sh` detects failures by checking for the presence of
"$REF_DIR/failed-plugins.txt". However failed-plugins.txt was not deleted
before the script starts, so errors in a prior run will cause all future
runs to fail. This causes problems with the jenkins helm chart: since
$REF_DIR points to an emptyDir volume, failed-plugins.txt persists until
the pod is deleted or moved to another node.