Fail fast while publishing variants
It's currently misleading: if an error occurs on a given variant, the whole
thing will continue happily.

This is probably nice because it will keep publishing other variants if
one has an issue, but it also certainly blurries what is going wrong
in the publishing process. So I think we should rather use `&&`.