Use Debian Buster instead of Debian Sid for jdk11
Debian Sid is the "unstable" version of Debian.  It is never "released"
but remains in a perpetual state of preparation for the next release.
The jdk11 docker image chose to use Debian Sid as their operating
system base.

That is reasonable for development, since it assures they are seeing
the most recent changes from the operating system development team.

It is not reasonable for production use.  The Jenkins base image should
use a production operating system version.

Debian Stretch (9) is released but does not provide a JDK 11 package.
Ubuntu 18.04 is released but provides a JDK 10 image instead of JDK 11.

Debian Buster (10) is expected to release in early 2019.  It already
provides a JDK 11 package based on OpenJDK 11.0.1.  It will likely be
released before we're ready to support an LTS version of Jenkins on
Java 11.