Sync Dockerfile-alpine with alpine branch
It was not immediately clear to me which branch is used to build which
image.  There are images for "standard", slim, and alpine.  There are
branches for master and alpine.  The Dockerfile-alpine on the master
branch does not seem to pass tests.  Here's hoping that the
Dockerfile-alpine on the alpine branch does better.