Use Debian Buster for JDK 11 Docker image
The docker image build fails now because openjdk:jdk-11 is based on Debian
Sid and there is not a version of tini 0.16.1 available for Debian Sid.

Debian Sid is the unstable Debian.  It is intentionally never released.
It does not have a security team.  It is the least stable of any of the
Debian versions.

Debian Buster will be Debian 10 when it is released in the first half
of 2019.  It already includes Java 11 as the default Java and is already
providing Java 11.0.1 as its default JDK.  Rather than be based on Sid
(that is never stable), let's base ourselves on Buster that is becoming
more and more stable.  Buster receives packages from Sid after specific
stability criteria are met.  It should generally be more stable than Sid.

Since tini is included in Debian Buster, this change also switches to
use tini as bundled with Debian Buster, instead of using a specific
pre-built version of tini.

Matching change is being proposed on the java11 branch.