Use updated OS packages at build time
Due to how significantly out of date the various base images get, this
adds an `apt-get upgrade` step to the various Dockerfiles to ensure that
updated operating system packages are included. This will also help
prevent this image from becoming too outdated when OS packages are
discovered to contain CVEs in the future.