Remove 'Publish Experimental' stage
7e270c32e43db9f7a1c039128fc2c7f0c61698f4 removed the publish-experimental
script.  That commit enabled parallel builds and moved the publish.sh
script into the .ci/ directory.

Unfortunately, it did not remove the reference to publish-experimental
from the Jenkinsfile or other files in the repository.

I assume the intent was to move all publishing to trusted.ci, including
both the publishing of experimental builds and official builds.

Even if that was not the intent, this does not make things any worse.
There is no make target for publish-experimental, so this will allow the
job to complete successfully, though the ci.jenkins.io job will now only
run tests and not publish any experimental images.