Replace Debian Stretch with Debian Buster & JDK 8u282
- Update Java from 8u242 to 8u282
- Move from Debian Stretch to Debian Buster

https://wiki.debian.org/DebianReleases notes that Debian Stretch end of
life was July 6, 2020.  See long term support end of life date on the
same page (currently estimated 2022).

The previous Docker JDK 8 base image (openjdk:8-jdk-Stretch) has not
been updated in a year.  This change brings in the most recent Java 8
release (8u282) and updates to a currently maintained version of the
Debian operating system (Buster).

Needs a blog post announcing the change at the same time as the first
weekly build that uses this Docker configuration.  Do not merge the pull
request until the matching blog post is available.