Use Git LFS 3.1.2 instead of 3.0.2
Only need to change docker-bake.hcl because it provides the arguments to docker bake

* https://github.com/git-lfs/git-lfs/releases/tag/v3.1.2
* https://github.com/git-lfs/git-lfs/releases/tag/v3.1.1

v3.1.2 changelog:

This is a bugfix release which fixes a bug in git lfs install and some issues
in our CI release processes, including one that prevented arm64 packages for
Debian 11 from being uploaded.

Bugs

* lfs: add old hook content to the list of old hooks #4878 (@bk2204)

Misc

* Revert "Merge pull request #4795 from bk2204/actions-checkout-v2" #4877 (@bk2204)
* .github/workflows: install packagecloud gem #4873 (@bk2204)

v3.1.1 changelog:

This is a bugfix release which fixes a syntax error in the release workflow.

Misc

* .github: fix syntax error in release workflow #4866 (@bk2204)

No changelog for v3.1.0, tag only due to issues in git LFS release process