Use git lfs 3.1.4, not git lfs 3.1.2
Updates git lfs even though we do not install git lfs on our Windows
Docker image for the controller and CVE-2022-24826 only affects Windows.
Better to run the latest version of git lfs than to explain why we're
not vulnerable to CVE-2022-24826.

https://www.cve.org/CVERecord?id=CVE-2022-24826 reports that

> On Windows, if Git LFS operates on a malicious repository with a `..exe`
> file as well as a file named `git.exe`, and `git.exe` is not found in
> `PATH`, the `..exe` program will be executed, permitting the attacker
> to execute arbitrary code.

Git LFS 3.1.3 changelog:

https://github.com/git-lfs/git-lfs/blob/032dca8ee69c193208cd050024c27e82e11aef81/CHANGELOG.md

Git LFS 3.1.4 changelog:

https://github.com/git-lfs/git-lfs/releases/tag/v3.1.4

This change intentionally updates all Dockerfile references, since there
is no reason to retain a default value that points to an outdated version
with a known CVE.  The Dockerfile updates should be unnecessary because
the argument is replaced by docker-bake.hcl, but making the change
everywhere retains consistency.