Identify release line based on Jenkins version
I previously made the mistake of attempting to reuse the logic that
was deciding whether this is the latest weekly or the latest LTS.
Reusing that logic fails to determine the correct release line for a
non-latest release.

If Jenkins 2.500 has already released but we need to build 2.499, then
LATEST_WEEKLY will be false and RELEASE_LINE should be "war".

If Jenkins 2.500.2 has already released but we need to build 2.500.1,
then LATEST_LTS will be false and RELEASE_LINE should be "war-stable".

The RELEASE_LINE calculation needs to be independent of the
calculation of latest releases, since they have different conditions.