Download from repo.jenkins-ci.org, not get.jenkins.io
The parallel build processes are much better served if the container
images can be created as soon as a release is published to the official
repository, repo.jenkins-ci.org, instead of waiting for the image to
be available on the mirrors.  The bandwidth used for container image
creation is small and the benefits of parallel creation of container
images are great.

Fixes https://github.com/jenkinsci/docker/issues/1671

Reverts PR https://github.com/jenkinsci/docker/pull/1648

Commits reverted include:

* a9f4d3126ecd - Provide RELEASE_LINE in all Windows test paths
* 3dfbe262cfd1 - Add RELEASE_LINE always in make.ps1
* 4a9dbf1f7f77 - Set RELEASE_LINE in Windows make.ps1
* a7b2df536759 - Limit RELEASE_LINE to use at build time
* 81e26f5d1af7 - Calculate RELEASE_LINE inside docker-bake.hcl
* 6f74cba3e589 - Export the RELEASE_LINE just like others
* 9b014edf320a - Identify release line based on Jenkins version
* d41891c6da16 - Use RELEASE_LINE instead of DOWNLOAD_DIR
* 7e57eb817225 - Include DOWNLOAD_DIR in build scripts