Remove install-plugins.sh from ubi9 Java 17 and Java 21 (#1817)
The install-plugins.sh stub is only included in Java 11 containers.
The stub announces to the user that it has been superseded by the plugin
installation manager tool.  Since it is not included in the other Java 17
and Java 21 containers, let's remove it from the single Java 17 container
and the single Java 21 container that include it.