Add Java 21 container for Windows
Change also includes an experiment to limit the tests that are run so
that we don't spend so much time in the Windows tests.

Goal: Save time and money by not duplicating slow tests on JDK versions

Many of the Windows tests run for a long time and don't tell us much
when we run them on all Java versions.  The change attempts to skip the
long running tests that are unlikely to detect a difference between JDK
versions.