Replace Java 21 previews with released versions
The arm/v7 support that was added as a preview has been dropped because
https://github.com/adoptium/adoptium-support/issues/962 notes that the
Eclipse Temurin project does not plan to deliver arm/v7 releases after
Java 17.

Includes released versions of Java 21 for s390x and ppc64le.

Uses the Eclipse Temurin container images for the JDK download because
released versions are available for all the supported platforms.