Check for Windows JDK releases, same as Linux (#1852)
* Check for Windows JDK releases, same as Linux

Returns the URL of the zip file, not the URL of the MSI file.  The Windows
container definitions do not depend on the results from this shell script,
so it is not an issue that it returns the zip file URL and not the MSI
file URL.

For example:

$ sh ./jdk-download-url.sh 21.0.3+9 windows "x64"
https://github.com/adoptium/temurin21-binaries/releases/download/jdk-21.0.3%2B9/OpenJDK21U-jdk_x64_windows_hotspot_21.0.3_9.zip

$ sh ./jdk-download-url.sh 17.0.11+9 windows "x64"
https://github.com/adoptium/temurin17-binaries/releases/download/jdk-17.0.11%2B9/OpenJDK17U-jdk_x64_windows_hotspot_17.0.11_9.zip

$ sh ./jdk-download-url.sh 11.0.23+9 windows "x64"
https://github.com/adoptium/temurin11-binaries/releases/download/jdk-11.0.23%2B9/OpenJDK11U-jdk_x64_windows_hotspot_11.0.23_9.zip

* Simplify updatecli JDK existence check for Windows

The Windows container images no longer depend on upstream containers
from Eclipse Temurin.  We use the Eclipse Temurin Windows MSI.