Pin the UBI 9 container 9.5-1734495538 (most recent release) (#1970)
* Use most recent UBI 9 container image

Dependabot should have detected this container image update, but it
did not.  Dependabot did not miss these updates in the platformlabeler
repository, where a simpler syntax is used for the test data containers.

I'm not sure if that is due to the AS clause or some other reason,
but we want to use the most recent UBI 9 container image as a base,
so this pull request makes sense in any case.

Latest image released 2 days ago from
https://catalog.redhat.com/software/containers/ubi9/ubi/615bcf606feffc5384e8452e

* Pin UBI 9.5-1734495538 as the UBI 9 container image

Most recent release from
https://catalog.redhat.com/software/containers/ubi9/ubi/615bcf606feffc5384e8452e