Revert "Bump Debian Bookworm Linux Version to 20250224"
Builds first started failing with this change:

* https://github.com/jenkinsci/docker/pull/1994

Failures are visible in this job:

* https://ci.jenkins.io/job/Packaging/job/docker/job/master/11859/

Failure message is:

 ERROR: failed to solve: failed to read expected number of bytes: unexpected EOF

Similar failure message was reported earlier in:

* https://github.com/jenkins-infra/helpdesk/issues/4551

In that case, it was related to a failure in the bug-mirror transparent
proxy for Docker.  If that is the case, then the tests will fail with
this revert as well.

This reverts commit 8f407b0b889baf7b766f5a171c4cae84a2fb7c20.